moved Alert to its own component and added its state to context
